refac: moved and renamed modules
