+++ cryptopals/s03/c23/solution_c23.py
@@ -2,7 +2,7 @@
 #   23 - Clone an MT19937 RNG from its output
 #
 
-from cryptopals.generator import MT19937
+from cryptopals.rng import MT19937
 
 
 class Cloner:
