+++ cryptopals/s04/c27/solution_c27.py
@@ -3,7 +3,7 @@
 #
 
 from cryptopals.oracle import Oracle
-from cryptopals.generator import Generator
+from cryptopals.utils import Generator
 from cryptopals.pkcs import PKCS7
 from cryptopals.symmetric import AES128CBC
 from cryptopals.solver import Detector
