+++ cryptopals/s05/c35/solution_c35.py
@@ -5,7 +5,7 @@
 from cryptopals.asymmetric import DiffieHellman
 from cryptopals.hash import SHA1
 from cryptopals.symmetric import AES128CBC
-from cryptopals.generator import Generator
+from cryptopals.utils import Generator
 from cryptopals.converter import Converter
