+++ cryptopals/s05/c37/server_c37.py
@@ -6,9 +6,8 @@
 
 from flask import Flask, request, jsonify, json
 from cryptopals.converter import Converter
-from cryptopals.generator import Generator
 from cryptopals.hash import SHA1
-from cryptopals.utils import Math
+from cryptopals.utils import Math, Generator
 
 global N, g, username, password
 global s, A, B, K, v, u, b, S
