+++ cryptopals/s05/c38/server_c38.py
@@ -6,8 +6,7 @@
 
 from flask import Flask, request, jsonify, json
 from cryptopals.converter import Converter
-from cryptopals.generator import Generator
-from cryptopals.utils import Math
+from cryptopals.utils import Math, Generator
 from cryptopals.hash import SHA1
 
 global N, g, username, P
